Renamed ThreadPool.h to threadpool.h
